+++ src/components/MatchProfilesForm/MatchProfilesForm.js
@@ -2,13 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { Field } from 'redux-form';
-import { identity } from 'lodash';
 
 import {
   Headline,
   TextArea,
   TextField,
-  Select,
   Accordion,
   AccordionSet,
 } from '@folio/stripes/components';
@@ -16,41 +14,23 @@ import stripesForm from '@folio/stripes/form';
 
 import { FullScreenForm } from '../FullScreenForm';
 import { validateRequiredField } from '../../utils';
-import { DATA_TYPES } from '../../utils/constants';
 
-const formName = 'jobProfilesForm';
-const dataTypes = DATA_TYPES.map(dataType => ({
-  value: dataType,
-  label: dataType,
-}));
+const formName = 'matchProfilesForm';
 
-export const JobProfilesFormComponent = props => {
+export const MatchProfilesFormComponent = props => {
   const {
     pristine,
     submitting,
-    initialValues,
     handleSubmit,
     onCancel,
   } = props;
 
-  const isEditMode = Boolean(initialValues.id);
   const isSubmitDisabled = pristine || submitting;
-
-  const paneTitle = isEditMode
-    ? (
-      <FormattedMessage id="ui-data-import.edit">
-        {txt => `${txt} ${initialValues.name}`}
-      </FormattedMessage>
-    )
-    : <FormattedMessage id="ui-data-import.settings.jobProfiles.newProfile" />;
-  const headLine = isEditMode
-    ? initialValues.name
-    : <FormattedMessage id="ui-data-import.settings.jobProfiles.newProfile" />;
 
   return (
     <FullScreenForm
-      id="job-profiles-form"
-      paneTitle={paneTitle}
+      id="match-profiles-form"
+      paneTitle={<FormattedMessage id="ui-data-import.settings.matchProfiles.newProfile" />}
       submitMessage={<FormattedMessage id="ui-data-import.save" />}
       isSubmitDisabled={isSubmitDisabled}
       onSubmit={handleSubmit}
@@ -61,7 +41,7 @@ export const JobProfilesFormComponent = props => {
         tag="h2"
         data-test-header-title
       >
-        {headLine}
+        <FormattedMessage id="ui-data-import.settings.matchProfiles.newProfile" />
       </Headline>
       <AccordionSet>
         <Accordion
@@ -77,22 +57,6 @@ export const JobProfilesFormComponent = props => {
               validate={[validateRequiredField]}
             />
           </div>
-          <div data-test-accepted-data-types-field>
-            <FormattedMessage id="ui-data-import.settings.jobProfiles.chooseAcceptedDataType">
-              {placeholder => (
-                <Field
-                  label={<FormattedMessage id="ui-data-import.settings.jobProfiles.acceptedDataType" />}
-                  name="dataType"
-                  component={Select}
-                  required
-                  itemToString={identity}
-                  validate={[validateRequiredField]}
-                  dataOptions={dataTypes}
-                  placeholder={placeholder}
-                />
-              )}
-            </FormattedMessage>
-          </div>
           <div data-test-description-field>
             <Field
               label={<FormattedMessage id="ui-data-import.description" />}
@@ -102,11 +66,11 @@ export const JobProfilesFormComponent = props => {
           </div>
         </Accordion>
         <Accordion
-          label={<FormattedMessage id="ui-data-import.settings.jobProfiles.overview" />}
+          label={<FormattedMessage id="ui-data-import.settings.matchProfiles.details" />}
           separator={false}
         >
           <div>
-            {/* will be implemented in https://issues.folio.org/browse/UIDATIMP-152 */}
+            {/* will be implemented in https://issues.folio.org/browse/UIDATIMP-175 */}
           </div>
         </Accordion>
       </AccordionSet>
@@ -114,16 +78,16 @@ export const JobProfilesFormComponent = props => {
   );
 };
 
-JobProfilesFormComponent.propTypes = {
-  initialValues: PropTypes.object.isRequired,
+MatchProfilesFormComponent.propTypes = {
   pristine: PropTypes.bool.isRequired,
   submitting: PropTypes.bool.isRequired,
   handleSubmit: PropTypes.func.isRequired,
   onCancel: PropTypes.func.isRequired,
 };
 
-export const JobProfilesForm = stripesForm({
+export const MatchProfilesForm = stripesForm({
   form: formName,
   navigationCheck: true,
   enableReinitialize: true,
-})(JobProfilesFormComponent);
+})(MatchProfilesFormComponent);
+