+++ test/bigtest/interactors/match-profiles.js
@@ -20,6 +20,7 @@ import { ActionMenuInteractor } from './action-menu-interactor';
   searchField = new TextFieldInteractor('#input-match-profiles-search');
   searchSubmitButton = new ButtonInteractor('[data-test-search-and-sort-submit]');
   searchSubmitButtonDisabled = property('[data-test-search-and-sort-submit]', 'disabled');
+  newMatchProfileButton = new ButtonInteractor('[data-test-new-button]');
 }
 
 export const matchProfiles = new MatchProfilesInteractor();